Compact blocks changes to support large blocks
This changes the data structures in compact blocks to make it capable
of handling larger blocks.

This bumps the upper boundry of the index range from 2^16 to 2^32.

Using names in BIP152:

Fields                                  | Old size  | New size
PrefilledTransaction::index             | 1,3 bytes | 1,3,5 bytes
HeaderAndShortIDs::shortids_length      | 1,3 bytes | 1,3,5 bytes
HeaderAndShortIDs::prefilledtxn_length  | 1,3 bytes | 1,3,5 bytes
BlockTransactionRequest::indexes_length | 1,3 bytes | 1,3,5 bytes
BlockTransactionRequest::indexes        | 1,3 bytes*indexes_length | 1,3,5 bytes*indexes_length
BlockTransactions::transactions_length  | 1,3 bytes | 1,3,5 bytes

+++ src/blockencodings.cpp
@@ -4,15 +4,15 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "blockencodings.h"
+#include "chainparams.h"
 #include "consensus/consensus.h"
 #include "consensus/validation.h"
-#include "chainparams.h"
 #include "hash.h"
+#include "main.h"
 #include "maxblocksize.h"
 #include "random.h"
 #include "streams.h"
 #include "txmempool.h"
-#include "main.h"
 #include "util.h"
 
 #include <unordered_map>
@@ -83,16 +83,16 @@ void validateCompactBlock(const CompactBlock& cmpctblock, uint64_t currMaxBlockS
 
     validateNumTxs(cmpctblock.shorttxids.size() + cmpctblock.prefilledtxn.size(), currMaxBlockSize);
 
-    int32_t lastprefilledindex = -1;
+    int64_t lastprefilledindex = -1;
     for (size_t i = 0; i < cmpctblock.prefilledtxn.size(); i++) {
         if (cmpctblock.prefilledtxn[i].tx.IsNull())
             throw std::invalid_argument("null tx in compact block");
 
-        lastprefilledindex += cmpctblock.prefilledtxn[i].index + 1; //index is a uint16_t, so cant overflow here
-        if (lastprefilledindex > std::numeric_limits<uint16_t>::max())
+        lastprefilledindex += static_cast<uint64_t>(cmpctblock.prefilledtxn[i].index) + 1; //index is a uint32_t, so cant overflow here
+        if (lastprefilledindex > std::numeric_limits<uint32_t>::max())
             throw std::invalid_argument("tx index overflows");
 
-        if (static_cast<uint32_t>(lastprefilledindex) > cmpctblock.shorttxids.size() + i) {
+        if (static_cast<uint64_t>(lastprefilledindex) > cmpctblock.shorttxids.size() + i) {
             // If we are inserting a tx at an index greater than our full list of shorttxids
             // plus the number of prefilled txn we've inserted, then we have txn for which we
             // have neither a prefilled txn or a shorttxid!
